feat: add some packages
