+++ src/ws_server/index.ts
@@ -1,0 +1,7 @@
+import { WebSocketServer } from 'ws';
+
+const WS_PORT = 3000;
+
+export const wsServer = new WebSocketServer({ port: WS_PORT });
+
+export default wsServer;
